Prevent hardlock caused by interacting with Cornifer in KE
It's possible to reach Cornifer without hitting the big title popup trigger when entering KE for the first time through a non-vanilla entrance to DE_03 (and quite likely with [right2]). This can lead to a hardlock.

I think having a fix specific to Deepnest_East_03 is sufficient, as anything generic would be unnecessarily intrusive.