Easy KP
A collection of logic changes that ensure that, for each location, the item, area and room logic are the same (with vanilla transitions).

In fact, Dash Slash and Basin Vessel have different logic, but this is due to the lemm access requirement (the nearby soul totems have the same absolute logic).

Settings: Easy KP, non cursed masks/notches/notch costs/nail.