Save Persistent bool data after the game does
Previously, when warping from a scene to itself, the PersistentBoolData would be saved as active (by rando, through the activeSceneChanged hook) and then the item from the old scene would save the PersistentBoolData as inactive (through the SaveLevelState hook).

For the most part, this was only a problem in transition rando.